+++ Runtime/MagicLeapLoader.cs
@@ -186,6 +186,10 @@ namespace UnityEngine.XR.MagicLeap
             StartSubsystem<XRInputSubsystem>();
             StartSubsystem<XRGestureSubsystem>();
 
+            // If we have the ControllerPose privilege, let's also ensure that the Controller Api is enabled
+            var hasControllerPose = MagicLeapPrivileges.IsPrivilegeApproved((uint) Privileges.ControllerPose);
+            inputSubsystem?.SetControllerApiEnabled(hasControllerPose);
+
             if (!isLegacyDeviceActive)
             {
                 StartSubsystem<XRDisplaySubsystem>();
@@ -206,6 +210,7 @@ namespace UnityEngine.XR.MagicLeap
 #endif // UNITY_2020_1_OR_NEWER
                 m_DisplaySubsystemRunning = true;
             }
+
             return true;
         }